Working with UrlRewriting as the first way of working with session management. Using parameters instead of attributes and using sendRedirect to SimpleSumSquare.java And printing error message when dividing by zero, using include instead of forward
1- Modified the way of taking number values from the SimpleArithmetic.java file, and instead of taking values
from parameters and setting it as an attribute on the HttpServletRequest object, I just send the request as it
is, because the parameters still exist (we are forwarding the same request), and then fetching parameters from
the responsible classes for the operation.

2- Used sendRedirect in SimpleAddition class instead of forwarding the request, so to experiment with facilities
of the method and to see the difference between it and RequestDispatcher Interface.

3- Added an HTML error message (included from server-side - not JS DOM) when trying to calculate a division by 0